pyproject.toml
[project]
name = "mlh-pe-hackathon"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.0",
    "flask>=3.1",
    "peewee>=3.17",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]
